+++ entrypoints/popup/main.tsx
@@ -1,3 +1,7 @@
 import { render } from 'preact';
+import { Popup } from './Popup';
+import './style.css';
 
-render(<div>NetCarve</div>, document.getElementById('root')!);
+const version = __NETCARVE_VERSION__;
+
+render(<Popup version={version} />, document.getElementById('root') as HTMLElement);